[sys, tests, cmake] Add projects to run tests from IDE.
Now, we rely on an agreement that successful tests are in the `pass` set,
and tests that may fail are in the `fail` set. We build projects to run and
initialize `pass` and `fail` separately. However, we clean them both
simultaneously. Note, that is possible to use other names for sets instead
of `pass` and `fail` (see `sapfor_add_test` CMake function for details).